app/pyproject.toml
[project]
name = "fotoscdo"
version = "0.1.0"
description = "Fotos CDO API (FastAPI + PostGIS + SFTP)"
requires-python = ">=3.11"
dependencies = [
  "fastapi~=0.115",
  "uvicorn[standard]~=0.30",
  "python-dotenv~=1.0",
  "pydantic-settings~=2.4",
  "SQLAlchemy~=2.0",
  "psycopg2-binary~=2.9",
  "geoalchemy2~=0.15",
  "shapely~=2.0",
  "pillow~=10.4",
  "piexif~=1.1",
  "paramiko~=3.4",
  "opentelemetry-sdk~=1.27",
  "opentelemetry-api~=1.27",
  "opentelemetry-instrumentation-fastapi~=0.48b0",
  "opentelemetry-exporter-otlp~=1.27",
]

[tool.uvicorn]
factory = false
host = "0.0.0.0"
port = 8000
reload = true